requirements.txt
flask
requests
faiss-cpu
numpy
gunicorn
